requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.0.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.1.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
python-dotenv==1.0.0
pytest>=7.4.0
httpx>=0.25.0
google-api-python-client>=2.80.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
